requirements.txt
dash==2.18.2
gspread==6.1.4
protobuf==5.29.2
python-dotenv==1.0.1
gunicorn
pandas
dash-tools
setuptools==68
wheel